Add check for vanilla Worn Armors.
NPC overrides that use a Worn Armor that is not actually an override
(i.e. the vanilla/DLC already used it) won't be counted as a body mod.

This was generally only an issue for non-human actors, which are already
being excluded by other rules, but it's still a possibility for humans.