Reduce visibility of body patcher.
A few people have found this and reported issues. It was written as a
quick fix for a narrow use case, and was never really intended to be
publicly discoverable or widely used.

Don't need to get rid of it entirely if some people are finding a use
for it, but should be hidden to reduce confusion and general issues.